pyproject.toml
[tool.poetry]
name = "portal"
version = "0.1.0"
description = ""
authors = ["Ansar Nurzhanov <nansarik@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
